backend transaction/transfer/bill pay worked

+++ banking-system-prototype/src/main/java/com/banking/prototype/repositories/AccountInfoRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AccountInfoRepository extends JpaRepository<AccountInfo, Long>{
 
-    AccountInfo getAccountInfoByUserId(int userId);
+    AccountInfo getAccountInfoByUserId(Long userId);
 }
