+++ banking-system-prototype/src/main/java/com/banking/prototype/repositories/BankAccountRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface BankAccountRepository extends JpaRepository<BankAccount, Long> {
-    List<BankAccount> findAllByInfo_UserId(int userId);
+    List<BankAccount> findAllByInfo_UserId(Long userId);
 }
